move lwjgl native library loading and logging config from graphics to timer
